Update Piece.java
#Piece Class v0.0.2  
#The TestMove Commit  
#Pull Request by KDerby123  
  
--------------------------------  
  
This pull request's main purpose is to introduce the abstract testMove() method into the Piece class. The testMove() method will be implemented in the Pawn, Bishop, Knight, Rook, Queen, and King classes, given that each piece moves differently. As a first test case, it will utilize the notSameColor() method (also implemented in the Piece class in this pull request) to filter whether the piece is of the same color.  
  
----------------------  
  
End Pull Request.